Removed unused files and refactored UI files
